Remove mapUtils from parser src since they weren't used
