+++ packages/v8-deopt-parser/test/traceMapsHelpers.js
@@ -1,7 +1,6 @@
 import * as path from "path";
 import { fileURLToPath } from "url";
 import { writeFile } from "fs/promises";
-import { edgeToString, getMapIdsFromICs } from "../src/mapUtils.js";
 
 // @ts-ignore
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
@@ -112,6 +111,22 @@ export async function writeMapSnapshot(logFileName, deoptInfo) {
 	const outFileName = logFileName.replace(".v8.log", ".mapTree.txt");
 	const outPath = path.join(__dirname, "snapshots", outFileName);
 	await writeFile(outPath, snapshot, "utf8");
+}
+
+/**
+ * @param {Iterable<import('..').ICEntry>} ics
+ * @returns {Set<string>}
+ */
+function getMapIdsFromICs(ics) {
+	/** @type {Set<string>} */
+	const mapIds = new Set();
+	for (const entry of ics) {
+		for (const update of entry.updates) {
+			mapIds.add(update.map);
+		}
+	}
+
+	return mapIds;
 }
 
 /**
@@ -242,6 +257,47 @@ function generateMapTree(
 	}
 
 	return output;
+}
+
+/**
+ * @param {import('../').MapEdge} edge
+ */
+function getEdgeSymbol(edge) {
+	switch (edge.subtype) {
+		case "Transition":
+			return "+";
+		case "Normalize": // FastToSlow
+			return "⊡";
+		case "SlowToFast":
+			return "⊛";
+		case "ReplaceDescriptors":
+			return edge.name ? "+" : "∥";
+		default:
+			return "";
+	}
+}
+
+/**
+ * @param {import('../').MapEdge} edge
+ */
+function edgeToString(edge) {
+	let s = getEdgeSymbol(edge);
+	switch (edge.subtype) {
+		case "Transition":
+			return s + edge.name;
+		case "SlowToFast":
+			return s + edge.reason;
+		case "CopyAsPrototype":
+			return s + "Copy as Prototype";
+		case "OptimizeAsPrototype":
+			return s + "Optimize as Prototype";
+		default:
+			if (edge.subtype == "ReplaceDescriptors" && edge.name) {
+				return edge.subtype + " " + getEdgeSymbol(edge) + edge.name;
+			} else {
+				return `${edge.subtype} ${edge?.reason ?? ""} ${edge?.name ?? ""}`;
+			}
+	}
 }
 
 /**
